Add missing type cast
Without a typecast combining uint8_t and int is ambigous (this was the
case with byte as type previously as well):

.../Arduino/libraries/SmartLCD/SmartLCDI2C.cpp: In member function ‘uint8_t SmartLCD::ReadADC()’:
.../Arduino/libraries/SmartLCD/SmartLCDI2C.cpp:530:25: warning: ISO C++ says that these are ambiguous, even though the worst conversion for the first is better than the worst conversion for the second:
  530 |         Wire.requestFrom(_I2C_Add, 2);
      |         ~~~~~~~~~~~~~~~~^~~~~~~~~~~~~
In file included from .../Arduino/libraries/SmartLCD/SmartLCDI2C.h:35,
                 from .../Arduino/libraries/SmartLCD/SmartLCDI2C.cpp:36:
.../arduino/sketchbook/arduino_base/arduino_ci/cpp/arduino/Wire.h:176:11: note: candidate 1: ‘uint8_t TwoWire::requestFrom(int, int)’
  176 |   uint8_t requestFrom(int address, int quantity) {
      |           ^~~~~~~~~~~
.../arduino/sketchbook/arduino_base/arduino_ci/cpp/arduino/Wire.h:172:11: note: candidate 2: ‘uint8_t TwoWire::requestFrom(uint8_t, uint8_t)’
  172 |   uint8_t requestFrom(uint8_t address, uint8_t quantity) {
      |           ^~~~~~~~~~~

While at it, change to rather cast the second integer argument.

+++ SmartLCDI2C.cpp
@@ -527,7 +527,7 @@ uint8_t SmartLCD::ReadADC(void)
 {
 	uint8_t Data = 0;
 
-	Wire.requestFrom(_I2C_Add, 2);
+	Wire.requestFrom(_I2C_Add, (uint8_t)2);
 	while (Wire.available())
 	{
 		Data = Wire.read();
@@ -543,7 +543,7 @@ uint8_t SmartLCD::Status(void)
 {
 	uint8_t Data = 0;
 
-	Wire.requestFrom((int)_I2C_Add, 1);
+	Wire.requestFrom(_I2C_Add, (uint8_t)1);
 	while (Wire.available())
 	{
 		Data = Wire.read();
